added custom UInt128 class to logic

+++ scripts/bit_tools.py
@@ -1,17 +1,19 @@
+from uint128 import UInt128
+
 # import itertools
 def as_mask(spin_indices: list) -> int:
-    mask = 0
+    mask = UInt128(0)
     for j in spin_indices:
-        mask |= 1 << j
+        mask |= UInt128(1) << j
     return mask
 
 
 def make_state(spin_indices: list, state: int):
     # state = n-bit integer
-    mask = 0
+    mask = UInt128(0)
     i = 0
     for spin_idx in spin_indices:
-        mask |= ((state & (1 << i)) >> i) << spin_idx
+        mask |= ((state & (UInt128(1) << i)) >> i) << spin_idx
         i += 1
     return mask
 
@@ -25,11 +27,10 @@ def make_state(spin_indices: list, state: int):
 #     return arr.reshape(-1, la)
 
 
-def bitperm(perm: list,  x: int):
+def bitperm(perm: list,  x: UInt128):
     # does the specified permutation on the bits of x
-    y = 0
+    y = UInt128(0)
     for i in range(len(perm)):
-        y |= ((x & (1<<i)) >> i) << perm[i]
+        y |= ((x & (UInt128(1)<<i)) >> i) << perm[i]
 
     return y
-
